fix(dhis2): fixed tests parameter field
