+++ src/workspaces/features/RunPipelineDialog/DHIS2Widget.generated.tsx
@@ -13,13 +13,12 @@ export type GetConnectionBySlugQueryVariables = Types.Exact<{
 }>;
 
 
-export type GetConnectionBySlugQuery = { __typename?: 'Query', connectionBySlug?: { __typename: 'CustomConnection' } | { __typename: 'DHIS2Connection', queryMetadata: { __typename?: 'DHIS2QueryResultPage', totalItems: number, error?: Types.Dhis2ConnectionError | null, items?: Array<{ __typename: 'DHIS2MetadataItem', id?: string | null, name: string } | { __typename: 'DHIS2OrganisationUnitLevel', id?: string | null, name: string, level?: number | null }> | null } } | { __typename: 'GCSConnection' } | { __typename: 'IASOConnection' } | { __typename: 'PostgreSQLConnection' } | { __typename: 'S3Connection' } | null };
+export type GetConnectionBySlugQuery = { __typename?: 'Query', connectionBySlug?: { __typename?: 'CustomConnection' } | { __typename?: 'DHIS2Connection', queryMetadata: { __typename?: 'DHIS2QueryResultPage', totalItems: number, error?: Types.Dhis2ConnectionError | null, items?: Array<{ __typename: 'DHIS2MetadataItem', id?: string | null, name: string } | { __typename: 'DHIS2OrganisationUnitLevel', id?: string | null, name: string, level?: number | null }> | null } } | { __typename?: 'GCSConnection' } | { __typename?: 'IASOConnection' } | { __typename?: 'PostgreSQLConnection' } | { __typename?: 'S3Connection' } | null };
 
 
 export const GetConnectionBySlugDocument = gql`
     query getConnectionBySlug($workspaceSlug: String!, $connectionSlug: String!, $type: String!, $search: String, $perPage: Int, $page: Int) {
   connectionBySlug(workspaceSlug: $workspaceSlug, connectionSlug: $connectionSlug) {
-    __typename
     ... on DHIS2Connection {
       queryMetadata(type: $type, search: $search, perPage: $perPage, page: $page) {
         items {
